# Faster `character_length()` string function for ASCII-only case